Decomposed model
