+++ packages/model/include/model/params.h
@@ -4,25 +4,11 @@
 #include "geom/pose.h"
 #include "geom/angle.h"
 #include "common/math.h"
+#include "model/shape.h"
 
 #include "yaml-cpp/yaml.h"
 
-#include <boost/assert.hpp>
-
 namespace truck::model {
-
-struct Shape {
-    Shape();
-    Shape(const YAML::Node& node);
-
-    double width;
-    double length;
-    double base_to_rear;
-    int circles_count;
-
-    double radius() const;
-    std::vector<geom::Vec2> getCircleDecomposition(const geom::Pose& ego_pose) const;
-};
 
 struct WheelBase {
     WheelBase(const YAML::Node& node);
