+++ packages/model/src/params.cpp
@@ -1,4 +1,6 @@
-#include <model/params.h>
+#include "model/params.h"
+
+#include <boost/assert.hpp>
 
 namespace truck::model {
 
@@ -18,35 +20,6 @@ Limits<double> toLimits(const YAML::Node& node) {
 }  // namespace
 
 using namespace geom::literals;
-
-Shape::Shape() {}
-
-Shape::Shape(const YAML::Node& node) :
-    width(node["width"].as<double>()),
-    length(node["length"].as<double>()),
-    base_to_rear(node["base_to_rear"].as<double>()),
-    circles_count(node["circles_count"].as<int>()) {
-    BOOST_VERIFY(width > 0);
-    BOOST_VERIFY(length > 0);
-    BOOST_VERIFY(base_to_rear > 0);
-    BOOST_VERIFY(length > base_to_rear);
-    BOOST_VERIFY(circles_count * 2 * radius() > length);
-}
-
-double Shape::radius() const { return width / 2; }
-
-std::vector<geom::Vec2> Shape::getCircleDecomposition(const geom::Pose& ego_pose) const {
-    std::vector<geom::Vec2> points;
-    const double pos_first = -base_to_rear + radius();
-    const double pos_step = (length - 2 * radius()) / (circles_count - 1);
-
-    for (int i = 0; i < circles_count; i++) {
-        double offset = pos_first + (i * pos_step);
-        points.push_back(ego_pose.pos + offset * ego_pose.dir);
-    }
-
-    return points;
-}
 
 WheelBase::WheelBase(const YAML::Node& node) :
     width(node["width"].as<double>()),
