+++ packages/model/test/main.cpp
@@ -5,11 +5,39 @@
 #include "geom/vector.h"
 #include "geom/test/equal_assert.h"
 
+#include <cmath>
+
 using namespace truck::model;
 
 TEST(Params, yaml_params) {
     // check config
     Params params("config/model.yaml");
+}
+
+TEST(Shape, rearPoseToShapePolygon) {
+    Shape shape;
+    shape.width = 1.5;
+    shape.length = 3;
+    truck::geom::Vec2 rear_center{2, 2};
+    truck::geom::AngleVec2 dir(truck::geom::Angle::fromDegrees(45));
+    truck::geom::Pose rear_pose{rear_center, dir};
+
+    const auto polygon = shape.rearPoseToShapePolygon(rear_pose);
+
+    constexpr double eps = 1e-9;
+
+    EXPECT_EQ(polygon.size(), 4);
+    const double d = 3 * std::sqrt(2) / 2;
+    const double ax = 2 - d / 4;
+    const double ay = 2 + d / 4;
+    ASSERT_GEOM_EQUAL(polygon[0].x, ax, eps);
+    ASSERT_GEOM_EQUAL(polygon[0].y, ay, eps);
+    ASSERT_GEOM_EQUAL(polygon[1].x, ax + d, eps);
+    ASSERT_GEOM_EQUAL(polygon[1].y, ay + d, eps);
+    ASSERT_GEOM_EQUAL(polygon[2].x, ay + d, eps);
+    ASSERT_GEOM_EQUAL(polygon[2].y, ax + d, eps);
+    ASSERT_GEOM_EQUAL(polygon[3].x, ay, eps);
+    ASSERT_GEOM_EQUAL(polygon[3].y, ax, eps);
 }
 
 TEST(Twist, angularVelocity) {
